Fix StickyTlb
