+++ platforms/f2/templates/sticky-tlb-template.ts
@@ -1,5 +1,4 @@
 import {
-	AdvertisementLabelHandler,
 	CloseButtonHelper,
 	CloseToHiddenButtonHandler,
 	DebugTransitionHandler,
@@ -33,7 +32,6 @@ export function registerStickyTlbTemplate(registry: TemplateRegistry): Observabl
 			initial: [
 				StickyTlbBootstrapHandler,
 				StickyTlbConfigHandler,
-				AdvertisementLabelHandler,
 				HideSmartBannerHandler,
 				DebugTransitionHandler,
 			],
