+++ platforms/shared/context/base/ucp-base-context.setup.ts
@@ -1,4 +1,4 @@
-import { context, Dictionary } from '@wikia/ad-engine';
+import { context } from '@wikia/ad-engine';
 import { Injectable } from '@wikia/dependency-injection';
 import { BaseContextSetup } from '../../setup/_base-context.setup';
 
@@ -31,18 +31,5 @@ export class UcpBaseContextSetup extends BaseContextSetup {
 			'userId',
 			(window.mw as any).config.get('wgTrackID') || (window.mw as any).config.get('wgUserId'),
 		);
-
-		const stickySlotsLines: Dictionary = this.instantConfig.get('icStickySlotLineItemIds');
-		if (stickySlotsLines && stickySlotsLines.length) {
-			context.set('templates.stickyTlb.lineItemIds', stickySlotsLines);
-
-			if (this.instantConfig.get('icHiViLeaderboardUnstickTimeout')) {
-				context.set('options.unstickHiViLeaderboardAfterTimeout', true);
-				context.set(
-					'options.unstickHiViLeaderboardTimeout',
-					this.instantConfig.get('icHiViLeaderboardUnstickTimeout'),
-				);
-			}
-		}
 	}
 }
