+++ platforms/shared/templates/handlers/sticky-tlb/sticky-tlb-blocking-handler.ts
@@ -29,14 +29,16 @@ export class StickyTlbBlockingHandler implements TemplateStateHandler {
 		const lines: string[] = context.get('templates.stickyTlb.lineItemIds') || [];
 
 		if (Array.isArray(lines)) {
-			return lines.some((line) => {
-				const [lineId, geo] = line.split(':', 2);
+			return lines
+				.map((line) => line.toString())
+				.some((line) => {
+					const [lineId, geo] = line.split(':', 2);
 
-				return (
-					lineId.toString() === this.adSlot.lineItemId.toString() &&
-					(!geo || utils.geoService.isProperGeo([geo]))
-				);
-			});
+					return (
+						lineId.toString() === this.adSlot.lineItemId.toString() &&
+						(!geo || utils.geoService.isProperGeo([geo]))
+					);
+				});
 		}
 		return false;
 	}
